Move proxies
